@@ -17,19 +17,20 @@
         }
     ],
     "require": {
-        "php": "~7.2",
+        "php": ">= 8.2",
         "ext-json": "*",
-        "symfony/process": "~3|~4|~5",
-        "jeremeamia/superclosure": "~2",
+        "symfony/process": "^6.3",
         "lstrojny/hmmmath": ">=0.5.0",
-        "guzzlehttp/guzzle": "^6.3",
-        "slim/slim": "^3.12",
-        "friendsofphp/php-cs-fixer": "^2.16",
-        "sensiolabs/security-checker": "^6.0"
+        "guzzlehttp/guzzle": "^7.8",
+        "slim/slim": "^4.12",
+        "friendsofphp/php-cs-fixer": "^3.16",
+        "charescape/serialize-closure": "^3.8",
+        "pimple/pimple": "^3.5",
+        "fig/http-message-util": "^1.1"
     },
     "require-dev": {
-        "internations/testing-component": "1.3.0",
-        "phpunit/phpunit": "^7"
+        "internations/testing-component": "1.4.0",
+        "phpunit/phpunit": "^9"
     },
     "autoload": {
         "psr-4": {
@@ -42,9 +43,6 @@
         }
     },
     "scripts": {
-        "auto-scripts": [
-            "@security-check"
-        ],
         "post-install-cmd": [
             "@auto-scripts"
         ],
@@ -57,8 +55,11 @@
         "lint:check": [
             "@lint --dry-run"
         ],
-        "security-check": [
-            "security-checker security:check"
-        ]
+        "tests":  "phpunit"
+    },
+    "config": {
+        "allow-plugins": {
+            "php-http/discovery": false
+        }
     }
 }
