@@ -6,22 +6,22 @@ use Psr\Http\Message\ServerRequestInterface as Request;
 
 class RequestStorage
 {
-    private $pid;
+    private int $pid;
 
-    private $directory;
+    private string $directory;
 
-    public function __construct($pid, $directory)
+    public function __construct(int $pid, string $directory)
     {
         $this->pid = $pid;
         $this->directory = $directory;
     }
 
-    public function store(Request $request, $name, $data)
+    public function store(Request $request, string $name, mixed $data) : void
     {
         file_put_contents($this->getFileName($request, $name), serialize($data));
     }
 
-    public function read(Request $request, $name)
+    public function read(Request $request, string $name) : mixed
     {
         $fileName = $this->getFileName($request, $name);
 
@@ -29,31 +29,29 @@ class RequestStorage
             return [];
         }
 
-        $r = Util::deserialize(file_get_contents($fileName));
-
-        return $r;
+        return Util::deserialize(file_get_contents($fileName));
     }
 
-    public function append(Request $request, $name, $data)
+    public function append(Request $request, string $name, mixed $data) : void
     {
         $list = $this->read($request, $name);
         $list[] = $data;
         $this->store($request, $name, $list);
     }
 
-    public function prepend(Request $request, $name, $data)
+    public function prepend(Request $request, string $name, mixed $data) : void
     {
         $list = $this->read($request, $name);
         array_unshift($list, $data);
         $this->store($request, $name, $list);
     }
 
-    private function getFileName(Request $request, $name)
+    private function getFileName(Request $request, string $name) : string
     {
         return $this->directory . $this->pid . '-' . $name . '-' . $request->getUri()->getPort();
     }
 
-    public function clear(Request $request, $name)
+    public function clear(Request $request, string $name) : void
     {
         $fileName = $this->getFileName($request, $name);
 
