@@ -26,7 +26,7 @@ class MockBuilderIntegrationTest extends TestCase
     /** @var Server */
     private $server;
 
-    public function setUp()
+    public function setUp() : void
     {
         $this->matches = new MatcherFactory();
         $this->builder = new MockBuilder($this->matches, new ExtractorFactory());
@@ -35,7 +35,7 @@ class MockBuilderIntegrationTest extends TestCase
         $this->server->clean();
     }
 
-    public function tearDown()
+    public function tearDown() : void
     {
         $this->server->stop();
     }
@@ -86,7 +86,7 @@ class MockBuilderIntegrationTest extends TestCase
 
         $this->assertSame('response body', (string) $client->post('/foo')->getBody());
 
-        $this->assertContains('CLOSURE MATCHER: POST /foo', $this->server->getErrorOutput());
+        $this->assertStringContainsString('CLOSURE MATCHER: POST /foo', $this->server->getErrorOutput());
     }
 
     public function testCreateTwoExpectationsAfterEachOther()
