@@ -10,22 +10,22 @@ class HttpMockPHPUnitIntegrationBasePathTest extends AbstractTestCase
 {
     use HttpMockTrait;
 
-    public static function setUpBeforeClass()
+    public static function setUpBeforeClass() : void
     {
         static::setUpHttpMockBeforeClass(null, null, '/custom-base-path');
     }
 
-    public static function tearDownAfterClass()
+    public static function tearDownAfterClass() : void
     {
         static::tearDownHttpMockAfterClass();
     }
 
-    public function setUp()
+    public function setUp() : void
     {
         $this->setUpHttpMock();
     }
 
-    public function tearDown()
+    public function tearDown() : void
     {
         $this->tearDownHttpMock();
     }
